Added functionality to delay attaching a volume
This change adds new functionality to fix an issue where test kitchen attempts to attach a volume to early while operations are still being performed. Some customized versions of Openstack such as VMWare Integrated Openstack (VIO) may mark a volume as active while it is still performing operations. This causes test kitchen to attach the volume to early which results in errors.

This change now offers the ability to delay attaching the volume for a specified amount of time